Customer-trade-finance/requirement.txt
streamlit
pandas
google-cloud-aiplatform
google-cloud-storage
python-dotenv